Rename PercentEncoder to PercentEncode
