+++ lib/domain/SyntaxSearch.js
@@ -3,7 +3,7 @@
 import SyntaxModel from './SyntaxModel';
 import ResultsPresenter from './ResultsPresenter';
 import { LanguageFilterMode } from './SyntaxModes';
-import { percentEncode } from '../util/PercentEncoder';
+import { percentEncode } from '../util/PercentEncode';
 import request from 'request';
 
 export default class SyntaxSearch extends SyntaxModel {
